requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.4
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
textdistance==4.5.0
tzdata==2023.3
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2
